\name{ranks.heatmap}
\alias{ranks.heatmap}
\title{Plot Random Effect Rank Distributions}
\description{
    Plots the rank distribution matrix for a random effect generated by
    \code{\link{mlmer}}.
}
\usage{
    ranks.heatmap(x, col="red")
}
\arguments{
    \item{x}{a matrix with the rank distribution for a random effect generated
             by \code{\link{mlmer}}.}
    \item{col}{the color corresponding to "higher-than-expected" counts.}
}
\section{Side Effects}{
    Generates a new plot.
}
\seealso{
    \code{\link{mlmer}}
}
